Added file where we can put in the measurements
These measurements are universal for almost every robot design, so
I put them in. There are several for the field itself, which we will
measure out.

Later on, we will add more measurements as we make the robot more
complex and add more in the way of tools and parts.