Modified include guard format to make it easier to read
In the previous format, each file had to make sure that it did not
include another file twice. Now, files will control themselves, and
only do things if they have not been included. This makes it easier
to understand for most C programmers because this follows the standard
method for include guards, whereas the previous one was made by me,
and as such not the standard format.

The functionality should remain the same, and the compile times will
remain fast.