Uses int to represent speed, not char
This is because `char` is implied to be used to represent an ASCII
character, so using it to represent a speed is not good practice.

An int on the NXT Brick is 2 bytes, so just 1 byte more than char.
That means that it is not a big deal as far as memory usage goes.