Using native int size
On the NXT Bricks, `int` and `short` are each 2 bytes, so there is no change in RAM usage. However, it is generally better to use `int` because it is the native int size, making it easier on the CPU so it doesn't have to make any changes.

This will either make no difference or make it slightly faster.