Avoid confusion of what findGoal() returns
findGoal() implicitly finds the goal, but what does it return? It's
not easy to be sure of that without looking at its code; it could
be the distance or the colour of the goal, or maybe just if the goal
is there at all. To avoid this confusion, findGoal() returns nothing
and the calculations are done with other functions.