Major reorganisation
The code should all work the same, but it is safe to assume that it
is now completely untested.

I tried to make the code a lot cleaner, and I like it a lot, but it's
going to be more complex from here on out. That said, I do like how
easy it is to find all the functions now that they're separated into
files.